Bug 33998: Installer and onboarding have incorrect Font Awesome asset links
This patch corrects the Font Awesome asset links in the web installer.
They were not updated during the upgrade to Font Awesome 6.

Unrelated: The patch also removes the inclusion of jQueryUI. It is not
used in the installation or onboarding process.

To test, apply the patch and run the web installer. The display of
installation steps should include square icons indicating which steps
are complete.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>