Bug 34042: Item search broken by FontAwesome upgrade
This patch corrects a small regression in one of the templates for
outputting item search results. The template lost the change made to it
by  Bug 33070: "Remove use of can_edit_items."

The error meant that item searching wasn't working at all.

To test, apply the patch and perform an item search in the staff
interface. Results should be returned as expected.

Signed-off-by: Phil Ringnalda <phil@chetcolibrary.org>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>