Bug 34124: Improve in-page navigation on table settings page
This patch adds links to the table settings page so that the user can
jump directly to a particular table within a page section.

The patch also changes the logic of when in-page links are displayed: I
think it makes sense not to show a "Jump to page" link when there is
only one page. The page will show a "Jump to table" link only if there
is only one table.

To test, apply the patch and go to Administration -> Table settings.

- Expand and check multiple sections.
- In the "Administration" section, there are multiple pages.
  - Check that the "Jump to page" links work correctly.
- In the "Catalog" section, there is only one page: "detail."
  - Check that the "Jump to table" links work correctly.

Signed-off-by: Sam Lau <samalau@gmail.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 24b2cd367f511e6f5f3dfb4bfb9b34d1e1ec8068)
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>