Bug 29825: (follow-up) Some style fixes and modal correction
This patch makes a few coding style changes to the new JavaScript and
updates the Bootstrap modal markup to be in line with current standards.

The previous patches lacked a trigger for the preview modal, so I added
that. Since the content of the modal uses spaces for formatting I added
monospace formatting to the modal body.

Signed-off-by: Phil Ringnalda <phil@chetcolibrary.org>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>