Bug 34270: Upgrade and prune jQueryUI assets in the staff interface
This patch updates jQueryUI assets in the staff interface with an
upgraded and trimmed version which includes only support for jQueryUI
autocomplete.

Global staff interface CSS has been modified to remove unused
jQueryUI-related styles and to accommodate the updated assets.

The patch also removes some unused jQueryUI image assets from the OPAC
templates.

To test, apply the patch and rebuild the staff interface CSS. Test the
"Check out" form in the header to confirm that it still looks correct
and works correctly.

Signed-off-by: Émily-Rose Francoeur <emily-rose.francoeur@inLibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>