Bug 34196: Improve styling of sidebar form on funds admin page
This patch makes a few markup changes to the sidebar filter form on the
funds administration page in order to make it more usable and more
consistent with other similar forms.

The patch removes the link for expanding and collapsing the sidebar
filter because it looked bad and was inconsistent with all similar
forms.

To test, apply the patch and go to Administration -> Funds. Confirm that
the sidebar form looks correct. Compare to, for example, the sidebar
form at Acquisitions -> Invoices.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>