Bug 34197: Group and label vendor contact settings
This patch changes the layout of vendor contact settings, both in the
entry form and in the detail view. Now acquisitions-specific contact
settings and serials-specific contact settings are grouped together and
labeled as such.

The patch also corrects some markup to make the display more consistent.
JS handling of some labels has been removed because the structure change
eliminated the need for a "for" attribute.

The patch also makes a minor unrelated change to the display of discount
percentages to ensure that a blank value doesn't display.

To test, apply the patch and go to Acquisitions -> New vendor.

- In the vendor entry form, in the "Contacts" section, you should see
  two groups of checkboxes, "Acquisitions options" and "Serials
  options." Confirm that the labels can be clicked to check checkboxes.
- Test the process of adding another contact and confirm that the
  additional entry fields work correctly.
  - Confirm that the "Primary acquisitions contact" and "Primary serials
    contact" checkboxes act like radio buttons between the two contacts:
    Checking the box under one contact should uncheck the same box under
    the other contact.
- Confirm that all your data is saved correctly.
- Test the page during the edit process as well.
- View your updated vendor and confirm that the detail view looks
  correct.

Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>