Bug 34322: Correct icon triggering more fund search options
This patch updates the budget/fund administration header search form so
that it uses the "fa-slider" Font Awesome icon instead of a folder icon.
This makes it consistent with other search forms which have a similar
set of additional options.

To test, apply the patch and go to Administration -> Budgets.

- The "Search funds" form should have the "fa-sliders" icon.
- Clicking it should trigger the expansion of the additional search
  options.

Signed-off-by: Sam Lau <samalau@gmail.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>